Add basic methods to Model domains

+++ src/Services/Ordering/Ordering.Domain/Models/OrderItem.cs
@@ -2,14 +2,13 @@
 
 public class OrderItem : Entity<OrderItemId>
 {
-    internal OrderItem(OrderId orderId, ProductId productId, int quantity, int price)
+    internal OrderItem(OrderId orderId, ProductId productId, int quantity, decimal price)
     {
         Id = OrderItemId.Of(Guid.NewGuid());
         OrderId = orderId;
         ProductId = productId;
         Quantity = quantity;
         Price = price;
-        
     }
 
     public OrderId OrderId { get; private set; } = default!;
